requirements.txt
Flask
Pillow
flask-cors
tensorflow
numpy
h5py
opencv-python